==> 0.1 Model Guidelines.R <==

#Our model should be able to:

#1. Distinguish between bookings and non-bookings
#2. Identify time-trends at the week, month, day level
#3. Identify price elasticities
#4. Rank Accounts, Device Type and Match Types by probability of order
#5. Identify high revenue bookings
